Add minimum path cost to priority queue insertion
The old version was mainly so slow because it was spending lots of time
considering various piece displacements nearby, because they are short
movements and come up first in the queue. Instead of ordering by the
path length *so far*, we should order the queue by the minimum path length
to the end state.

I guess this is now more A* than Dijkstra?